feat(colorbar): basic colorbar example + fixed colobar size
